First Exercise with Array
In this exercise an user will have the opportunity to write a name in a
prompt. There’s a list with some names and if the name which the user
will choose is already in the list, it will not change anything,
otherwise, the name will be inserted in the list.